requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monai==1.4.0
mpmath==1.3.0
networkx==3.3
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.8
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
simpleitk==2.5.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
torch==2.3.1+cu121
torchio==0.20.8
torchvision==0.18.1+cu121
tqdm==4.67.1
triton==2.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.0
wrapt==1.17.2
